refactor(hub): move with-shape/classified → shape/via-classified (#629)
Pure move (grammar correction, no behavior change) — precedent: the #623
with-shape/money and with-shape/i18n moves. Every kernel/test/tsup/knip
specifier updated to the new path; strategy.ts's real ClassifiedStrategy/
NO_CLASSIFIED content moves to port/with/classified-strategy.ts (the
i18n-strategy precedent), leaving a re-export shim behind so existing
importers keep working. config-drift.ts's `dek` param switches from the
EnclaveKey alias to the bare CryptoKey global (identical type) and its two
persisted-schemas dynamic imports are rebased — both changes forced by the
new shape/via-classified/ location falling under the via-enclave-isolation
architecture guard (kernel/enclave/ imports banned there, even type-only,
with an empty allowlist). Two of config-drift.ts's dynamic imports and two
kernel-adjacent value imports (resolveClassifiedFields/guardClassifiedCompat
+ enforceClassifiedWrite/ClassifiedRevealError/ClassifiedVerifyError) are
temporarily grandfathered in VIA_SHAPE_ALLOWLIST — the classified binding is
still dormant (Task 6 compiles it in and retires the grandfather, per the
plan's own "grandfather retirements" acceptance criterion for that task).

Full classified suite (35 files / 168 tests), full hub suite (447 files /
3523 tests), typecheck, lint, and check:architecture all green.
