+++ knip.json
@@ -18,7 +18,7 @@
         "src/with-fork/snapshots/index.ts",
         "src/with-party/session/index.ts",
         "src/with-party/team/index.ts",
-        "src/with-shape/classified/index.ts",
+        "src/shape/via-classified/index.ts",
         "src/with-shape/satellites/index.ts"
       ]
     }
