+++ packages/hub/__tests__/classified/find-strategy-seam.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect } from 'vitest'
-import { NO_CLASSIFIED } from '../../src/with-shape/classified/strategy.js'
-import type { ClassifiedVerifyCtx } from '../../src/with-shape/classified/strategy.js'
-import { withClassified } from '../../src/with-shape/classified/active.js'
-import { classified } from '../../src/with-shape/classified/presets.js'
+import { NO_CLASSIFIED } from '../../src/shape/via-classified/strategy.js'
+import type { ClassifiedVerifyCtx } from '../../src/shape/via-classified/strategy.js'
+import { withClassified } from '../../src/shape/via-classified/active.js'
+import { classified } from '../../src/shape/via-classified/presets.js'
 import { ClassifiedNotEnabledError } from '../../src/kernel/errors.js'
 import { generateDEK } from '../../src/kernel/enclave/index.js'
 import { normalizeForVerify } from '../../src/kernel/enclave/classify/normalize.js'
