+++ packages/hub/__tests__/classified/forget-vdig.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect } from 'vitest'
 import { createNoydb } from '../../src/kernel/noydb.js'
 import { inlineMemory } from './harness.js'
 import { buildTombstone } from '../../src/kernel/enclave/index.js'
-import type { ClassifiedFieldSpec } from '../../src/with-shape/classified/descriptor.js'
+import type { ClassifiedFieldSpec } from '../../src/shape/via-classified/descriptor.js'
 
 const passwordSpec: ClassifiedFieldSpec = {
   _noydbClassified: true, preset: 'password', storage: 'digest-only',
