+++ packages/hub/__tests__/classified/stage2-spine.test.ts
@@ -5,8 +5,8 @@ import {
 } from '../../src/kernel/errors.js'
 import {
   ClassifiedConfigError as ShimConfig, ClassifiedRevealError as ShimReveal,
-} from '../../src/with-shape/classified/errors.js'
-import type { ClassifiedFieldSpec } from '../../src/with-shape/classified/descriptor.js'
+} from '../../src/shape/via-classified/errors.js'
+import type { ClassifiedFieldSpec } from '../../src/shape/via-classified/descriptor.js'
 
 describe('stage-2 spine', () => {
   it('EncryptedEnvelope accepts a _vdig ciphertext map', () => {
