+++ packages/hub/src/kernel/collection-config.ts
@@ -43,9 +43,9 @@ import type { IndexStrategy } from '../with-lookup/indexing/strategy.js'
 import type { IndexDef } from '../with-lookup/indexing/eager-indexes.js'
 import type { I18nTextDescriptor, DictKeyDescriptor, StaticDictDescriptor, DictionaryHandle } from '../port/with/i18n-strategy.js'
 import type { ComputedFields } from '../with-formula/computed/index.js'
-import { resolveClassifiedFields, ClassifiedConfigError, type ClassifiedEntry, type ResolvedClassified } from '../with-shape/classified/resolve.js'
-import { guardClassifiedCompat, type ClassifiedGuardCtx } from '../with-shape/classified/guards.js'
-import { NO_CLASSIFIED, type ClassifiedStrategy } from '../with-shape/classified/strategy.js'
+import { resolveClassifiedFields, ClassifiedConfigError, type ClassifiedEntry, type ResolvedClassified } from '../shape/via-classified/resolve.js'
+import { guardClassifiedCompat, type ClassifiedGuardCtx } from '../shape/via-classified/guards.js'
+import { NO_CLASSIFIED, type ClassifiedStrategy } from '../port/with/classified-strategy.js'
 import type { FieldMeta } from '../with-shape/introspection/field-meta.js'
 import type { CollectionMeta } from '../with-shape/introspection/meta.js'
 import type { RefDescriptor } from './refs.js'
