+++ packages/hub/src/kernel/errors.ts
@@ -2793,8 +2793,8 @@ export class EnclaveNotSupportedError extends NoydbError {
  * Raised when a collection's `classifiedFields` configuration is invalid
  * (e.g. a claimed field name collides with a rider companion or another
  * classified field). Homed in `kernel/errors.ts` (rather than
- * `with-shape/classified/errors.ts`) so `kernel/enclave/classify/*` can throw
- * it without importing with-*; `with-shape/classified/errors.ts` re-exports
+ * `shape/via-classified/errors.ts`) so `kernel/enclave/classify/*` can throw
+ * it without importing with-*; `shape/via-classified/errors.ts` re-exports
  * it under the same name for backward-compatible import paths.
  */
 export class ClassifiedConfigError extends Error {
