+++ packages/hub/src/kernel/types.ts
@@ -39,7 +39,7 @@ import type { HistoryStrategy } from '../with-commit/history/strategy.js'
 import type { ForgetStrategy } from '../with-audit/forget/strategy.js'
 import type { SnapshotStrategy } from '../with-fork/snapshots/strategy.js'
 import type { AttestationStrategy } from '../with-audit/attestation/strategy.js'
-import type { ClassifiedStrategy } from '../with-shape/classified/strategy.js'
+import type { ClassifiedStrategy } from '../port/with/classified-strategy.js'
 import type { TiersStrategy } from '../with-audit/tiers/strategy.js'
 import type { SealedRecordStrategy } from '../with-audit/sealed-record/strategy.js'
 import type { PortabilityStrategy } from '../with-audit/portability/strategy.js'
