+++ packages/hub/src/shape/via-classified/guards.ts
