+++ packages/hub/tsup.config.ts
@@ -46,7 +46,7 @@ const ENTRIES = {
   'sync/index': 'src/with-party/sync/index.ts',
   'util/index': 'src/kernel/util/index.ts',
   'attestation/index': 'src/with-audit/attestation/index.ts',
-  'classified/index': 'src/with-shape/classified/index.ts',
+  'classified/index': 'src/shape/via-classified/index.ts',
   'satellites/index': 'src/with-shape/satellites/index.ts',
   'tiers/index': 'src/with-audit/tiers/index.ts',
   'portability/index': 'src/with-audit/portability/index.ts',
